refactorizacion del codigo usando customHooks y agregada descripcion tecnica del proyecto en el archivo .md
