+++ src/api/configApi.js
@@ -1,5 +1,5 @@
 const configApi = {
-    apiBaseUrl: 'http://localhost:3000',
+    apiBaseUrl: import.meta.env.VITE_API_BASE_URL || 'http://localhost:3000', // Fallback value
     endpoints: {
         searchWorkerById: '/attendance',
         getAttendance: '/attendance',
