+++ src/components/ProtectedRoute.jsx
@@ -4,7 +4,7 @@ import { useUser } from "../context/userContext";
 const ProtectedRoute = () => {
     const { user } = useUser();
     const userLocal = localStorage.getItem("user"); 
-    console.log(user);
+    // console.log(user);
     return user || userLocal ? <Outlet /> : <Navigate to="/" />;
     }
