--- src/data/data.js
@@ -1,159 +1,0 @@
-const data = [
-  {
-    workers: [
-      {
-        id: 1,
-        cedula: "27451286",
-        nombre_completo: "John Doe",
-        departamento: "Ventas",
-        cargo: "Gerente",
-        fecha:"21/06/2024",
-        hora_entrada: "09:30 AM",
-        hora_salida: "06:00 PM",
-        genero: "Masculino",
-        direccion: "Calle 123",
-        telefono: "04121234567",
-        correo: "example123@gmail.com",
-        estatus: "Activo",
-      },
-      {
-        id: 2,
-        cedula: "27451287",
-        nombre_completo: "Jane Doe",
-        departamento: "Ventas",
-        cargo: "Vendedor",
-        fecha:"21/06/2024",
-        hora_entrada: "09:30 AM",
-        hora_salida: "06:00 PM",
-        genero: "Femenino",
-        direccion: "Calle 123",
-        telefono: "04121234567",
-        correo: "correocorreo@gmail.com",
-        estatus: "Activo",
-      },
-      {
-        id: 3,
-        cedula: "27451288",
-        nombre_completo: "Carlos Pérez",
-        departamento: "Marketing",
-        cargo: "Analista",
-        fecha:"21/06/2024",
-        hora_entrada: "08:00 AM",
-        hora_salida: "05:00 PM",
-        genero: "Masculino",
-        direccion: "Avenida 456",
-        telefono: "04121234568",
-        correo: "carlosperez@gmail.com",
-        estatus: "Activo",
-      },
-      {
-        id: 4,
-        cedula: "27451289",
-        nombre_completo: "Ana García",
-        departamento: "Recursos Humanos",
-        cargo: "Coordinadora",
-        fecha:"21/06/2024",
-        hora_entrada: "08:30 AM",
-        hora_salida: "05:30 PM",
-        genero: "Femenino",
-        direccion: "Calle 789",
-        telefono: "04121234569",
-        correo: "anagarcia@gmail.com",
-        estatus: "Activo",
-      },
-      {
-        id: 5,
-        cedula: "27451290",
-        nombre_completo: "Luis Rodríguez",
-        departamento: "IT",
-        cargo: "Desarrollador",
-        fecha:"21/06/2024",
-        hora_entrada: "09:00 AM",
-        hora_salida: "06:00 PM",
-        genero: "Masculino",
-        direccion: "Avenida 101",
-        telefono: "04121234570",
-        correo: "luisrodriguez@gmail.com",
-        estatus: "Activo",
-      },
-      {
-        id: 6,
-        cedula: "27451291",
-        nombre_completo: "María Fernández",
-        departamento: "Finanzas",
-        cargo: "Contadora",
-        fecha:"21/06/2024",
-        hora_entrada: "08:00 AM",
-        hora_salida: "04:30 PM",
-        genero: "Femenino",
-        direccion: "Calle 111",
-        telefono: "04121234571",
-        correo: "mariafernandez@gmail.com",
-        estatus: "Activo",
-      },
-      {
-        id: 7,
-        cedula: "27451292",
-        nombre_completo: "Pedro López",
-        departamento: "Ventas",
-        cargo: "Asistente",
-        fecha:"21/06/2024",
-        hora_entrada: "09:30 AM",
-        hora_salida: "06:00 PM",
-        genero: "Masculino",
-        direccion: "Calle 222",
-        telefono: "04121234572",
-        correo: "pedrolopez@gmail.com",
-        estatus: "Inactivo",
-      },
-      {
-        id: 8,
-        cedula: "27451293",
-        nombre_completo: "Lucía Torres",
-        departamento: "Marketing",
-        cargo: "Diseñadora",
-        fecha:"21/06/2024",
-        hora_entrada: "08:30 AM",
-        hora_salida: "05:30 PM",
-        genero: "Femenino",
-        direccion: "Avenida 333",
-        telefono: "04121234573",
-        correo: "luciatorres@gmail.com",
-        estatus: "Activo",
-      },
-      {
-        id: 9,
-        cedula: "27451294",
-        nombre_completo: "Jorge Martínez",
-        departamento: "IT",
-        cargo: "Administrador de Sistemas",
-        fecha:"21/06/2024",
-        hora_entrada: "09:00 AM",
-        hora_salida: "06:00 PM",
-        genero: "Masculino",
-        direccion: "Calle 444",
-        telefono: "04121234574",
-        correo: "jorgemartinez@gmail.com",
-        estatus: "Activo",
-      },
-      {
-        id: 10,
-        cedula: "27451295",
-        nombre_completo: "Laura Suárez",
-        departamento: "Finanzas",
-        cargo: "Analista Financiero",
-        fecha:"21/06/2024",
-        hora_entrada: "08:00 AM",
-        hora_salida: "04:30 PM",
-        genero: "Femenino",
-        direccion: "Avenida 555",
-        telefono: "04121234575",
-        correo: "laurasuarez@gmail.com",
-        estatus: "Activo",
-      },
-    ],
-    date: new Date().toLocaleDateString(),
-  },
-];
-
-export default data;
