+++ src/pages/PageHome.jsx
@@ -1,4 +1,4 @@
-import { useEffect, useState } from "react";
+import { useState } from "react";
 import Button from "../components/Button";
 import ModalLogin from "../components/ModalLogin";
 import { Header } from "../components/Header";
@@ -6,25 +6,27 @@ import WorkerDetails from "../components/WorkerDetails";
 import { Alert } from "@mui/material";
 import CheckIcon from "@mui/icons-material/Check";
 import { date } from "../utils/date";
-import configApi from "../api/configApi";
+
 import { useUser } from "../context/userContext";
+import useWorker from "../hooks/useWorker";
 
 export const PageHome = () => {
   const [showModal, setShowModal] = useState(false);
-  const [worker, setWorker] = useState(null);
   const [searchWorkerBtn, setSearchWorkerBtn] = useState(false);
   const [checkInBtn, setCheckInBtn] = useState(false);
   const [checkOutBtn, setCheckOutBtn] = useState(false);
   const [workerIdentity, setWorkerIdentity] = useState("");
-  const [checkInSuccess, setCheckInSuccess] = useState(false);
-  const [checkOutSuccess, setCheckOutSuccess] = useState(false);
-  const [loader, setLoader] = useState(false);
-  const {setErrorCredentials } = useUser();
+  const { setErrorCredentials } = useUser();
 
-  const handleSearchWorker = (e) => {
-    e.preventDefault();
-    setSearchWorkerBtn(true);
-  };
+
+  const { worker, loader, checkInSuccess, checkOutSuccess, handleSearchWorker} = useWorker(
+    workerIdentity,
+    searchWorkerBtn,
+    checkInBtn,
+    checkOutBtn,
+    setSearchWorkerBtn
+  );
+
 
   const handleCheckInBtn = () => {
     setCheckInBtn(true);
@@ -42,124 +44,6 @@ export const PageHome = () => {
     setErrorCredentials(null);
   };
 
-  const apiEndPointGetWorkerByid =
-    configApi.apiBaseUrl + configApi.endpoints.searchWorkerById;
-  const apiEndPointGetAttendance =
-    configApi.apiBaseUrl + configApi.endpoints.getAttendance;
-
-  // ------Buscar trabajador en la base de datos
-  useEffect(() => {
-    if (workerIdentity) {
-      if (searchWorkerBtn || checkInBtn || checkOutBtn) {
-        setLoader(true);
-        fetch(`${apiEndPointGetWorkerByid}/${workerIdentity}`, {
-          method: "GET",
-          headers: {
-            "Content-Type": "application/json",
-          },
-        })
-          .then((response) => {
-            if (response.ok) {
-              response.json().then((data) => {
-                setWorker(data.data[0]);
-                console.log(data);
-              });
-            }
-          })
-          .catch((error) => {
-            console.error("Error:", error);
-          })
-          .finally(() => {
-            setSearchWorkerBtn(false);
-            setLoader(false);
-          });
-      }
-    }
-  }, [searchWorkerBtn, workerIdentity]);
-
-  // ------Marcar hora de entrada en la base de datos
-  useEffect(() => {
-    if (checkInBtn){
-    fetch(apiEndPointGetAttendance, {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({ id: worker?.worker_id }),
-    })
-      .then((response) => {
-        if (response.ok) {
-          response.json().then((data) => {
-            setCheckInSuccess(true);
-            setWorker({
-              ...worker,
-              date_attendance_string: date,
-              date_attendance: date,
-              check_in_string: data.data.check_in_string,
-            });
-            console.log(data);
-          });
-        }
-      })
-      .catch((error) => {
-        console.error("Error:", error);
-      })
-      .finally(() => {
-        setCheckInBtn(false);
-      });
-  }
-}, [checkInBtn]);
-
-  // --------- Actualización de hora de salida en la base de datos
-  useEffect(() => {
-    if (!checkOutBtn) return;
-    if (worker.check_out) return;
-    fetch(apiEndPointGetAttendance, {
-      method: "PATCH",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({ id: worker.worker_id }),
-    })
-      .then((response) => {
-        if (response.ok) {
-          response.json().then((data) => {
-            setCheckOutSuccess(true);
-            setWorker({
-              ...worker,
-              check_out_string: data.data.check_out_string,
-            });
-            console.log(data);
-          });
-        }
-      })
-      .catch((error) => {
-        console.error("Error:", error);
-      })
-      .finally(() => {
-        setCheckOutBtn(false);
-      });
-  }, [checkOutBtn]);
-
-  useEffect(() => {
-    if (checkInSuccess) {
-      const timer = setTimeout(() => {
-        setCheckInSuccess(false);
-      }, 5000);
-      return () => clearTimeout(timer);
-    }
-  }, [checkInSuccess]);
-
-  useEffect(() => {
-    if (checkOutSuccess) {
-      const timer = setTimeout(() => {
-        setCheckOutSuccess(false);
-      }, 5000);
-      return () => clearTimeout(timer);
-    }
-  }, [checkOutSuccess]);
-
-  // console.log(worker);
 
   return (
     <>
