+++ src/utils/date.js
@@ -1,4 +1,4 @@
-const date = new Date().toLocaleDateString('en-GB', {
+const date = new Date().toLocaleDateString('es-VE', {
   day: '2-digit',
   month: '2-digit',
   year: 'numeric',
